# perf(osmoutils): use make([]T, size) hint in *SubDecCoinArrays