This change uses make([]T, size) when creating the slices that hold the final results from the subtractions, along with inserting values along with indices.

This results in these benchmarks

```shell
$ benchstat before.txt after.txt
name                    old time/op    new time/op    delta
SafeSubDecCoinArrays-8    4.23µs ± 6%    4.13µs ± 7%    ~     (p=0.243 n=9+10)
SubDecCoinArrays-8        4.09µs ± 4%    4.02µs ± 4%    ~     (p=0.052 n=10+10)

name                    old alloc/op   new alloc/op   delta
SafeSubDecCoinArrays-8    3.08kB ± 0%    3.06kB ± 0%  -0.78%  (p=0.000 n=10+10)
SubDecCoinArrays-8        3.08kB ± 0%    3.08kB ± 0%    ~     (all equal)

name                    old allocs/op  new allocs/op  delta
SafeSubDecCoinArrays-8      75.0 ± 0%      74.0 ± 0%  -1.33%  (p=0.000 n=10+10)
SubDecCoinArrays-8          75.0 ± 0%      75.0 ± 0%    ~     (all equal)
```

and this very simple change over time aids reduce RAM/CPU cycles given that Osmosis is very popular.

Fixes #7756

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit


- **Performance Improvements**
	- Enhanced the efficiency of decimal coin array subtraction operations in the `osmoutils` package.
- **Tests**
	- Introduced benchmark tests for decimal coin array subtraction functions.
- **Refactor**
	- Optimized the initialization process of arrays in the `SubDecCoinArrays` and `SafeSubDecCoinArrays` functions.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->